Revert "Add website routing rules api & tests & example"
This reverts commit 6e46344f84203d9f32cac227c20e4892442c4617.
